+++ SDF/src/Structures/Mesh/Octree.h
@@ -13,13 +13,13 @@ namespace MeshStructures
 	public class Octree
 	{
 	public:
-		Octree(const int dep, const double siz, Vector4 ori, Octree* par = NULL);
+		Octree(const int dep, const float siz, Vector4 ori, Octree* par = NULL);
 		~Octree();
 
 		void Build(Face** tria, unsigned int length);
 		int GetCode(const Vector4 pt);
 		bool isLeaf;
-		void GetBoundary(double &siz, double &x, double &y, double &z);
+		void GetBoundary(float &siz, float &x, float &y, float &z);
 		void DrawOctree(bool recursive);
 
 		Octree* parent;						// otec, NULL ak sme hlavny vrchol
@@ -28,7 +28,7 @@ namespace MeshStructures
 		unsigned int count;					// pocet trojuholnikov, hranica pre robenie synov je min_count
 	private:
 		int depth;							// ako hlboko sme v octree, max je max_depth
-		double size;						// hranica kocky
+		float size;						// hranica kocky
 		Vector4 origin;						// stred kocky
 	};
 }
